+++ requirements.txt
@@ -1,5 +1,4 @@
 torch >= 1.0.0
-tensorflow >= 1.7.0
 funcsigs >= 1.0.2
 numpy >= 1.15.4
 mypy_extensions >= 0.4.1
