+++ texar/custom/__init__.py
@@ -4,4 +4,3 @@ Custom modules in Texar
 
 from texar.custom.activation import *
 from texar.custom.initializers import *
-from texar.custom.optimizers import *
