+++ texar/modules/pretrained/bert_utils_test.py
@@ -5,7 +5,7 @@ Unit tests for bert utils.
 import os
 import unittest
 
-from texar.modules.berts.bert_utils import *
+from texar.modules.pretrained.bert_utils import *
 
 
 class BertUtilsTest(unittest.TestCase):
